requirements.txt
boto3==1.18.30
botocore==1.21.30
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.0.1
Flask==2.0.1
Flask-Cors==4.0.0
flask-lambda-support==0.1.5
idna==3.4
importlib-metadata==4.7.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.31.0
s3transfer==0.5.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
